Improved the HealthScript
Improved the HealthScript. It first checks if the script is attached to player, since that influences how the script will behave. When the object the script is attached to runs out of health, it takes the appropriate action depending on if "isPlayer" is true or false. Finally, it updates a healthbar, but only if the script is attached to the player.